requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
duckdb>=0.7.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
lxml>=4.9.0
pytest>=7.0.0
pytest-mock>=3.10.0 
dotenv
